Re-organize 'HashStore' structure by moving 'HashAddress' out of filehashstore directory and refactor affected classes
